backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.1
python-dotenv
